int data type params for GDAL memraster creation

+++ qgis_plugin/mtf_estimator/mtf_estimator.py
@@ -286,15 +286,15 @@ class MtfEstimator:
         M = np.linalg.inv(m)
         col_list, row_list = np.matmul(M,(E-A).T)
         pxoffset = 5
-        col_min = np.int64(np.max([np.floor(np.min(col_list)) - pxoffset,1]))
-        col_max = np.int64(np.min([np.ceil(np.max(col_list))+pxoffset, xsize-1]))
-        row_min = np.int64(np.max([np.floor(np.min(row_list)) - pxoffset,1]))
-        row_max = np.int64(np.min([np.ceil(np.max(row_list))+pxoffset, ysize-1]))
+        col_min = int(np.max([np.floor(np.min(col_list)) - pxoffset,1]))
+        col_max = int(np.min([np.ceil(np.max(col_list))+pxoffset, xsize-1]))
+        row_min = int(np.max([np.floor(np.min(row_list)) - pxoffset,1]))
+        row_max = int(np.min([np.ceil(np.max(row_list))+pxoffset, ysize-1]))
         sub_gt = gt
         sub_gt[0] = gt[0] + gt[1]*col_min + gt[2]*row_min
         sub_gt[3] = gt[3] + gt[4]*col_min + gt[5]*row_min
-        sub_xsize = col_max-col_min
-        sub_ysize = row_max-row_min        
+        sub_xsize = int(col_max-col_min)
+        sub_ysize = int(row_max-row_min)
         
         memraster_drv = gdal.GetDriverByName('MEM')
         memraster = memraster_drv.Create('', sub_xsize, sub_ysize, 1, band.DataType)
